Refactor `reactCompilerLoader` to be consolidated into `ReactCompilerPlugin`
